Avoid harmless but unclean bounds checking patterns
Using (size + x > bound) to check bounds is prone to overflow issues. It
certainly shouldn't have ever appeared in protocol code. In this case,
it's harmless because the input is bounded to uint16, and because it's
later truncated to the size of the buffer.

This entire pile of code is due to be removed along with a new protocol,
using better serialization, but it's still worth fixing for hygiene.