Emit the contactAdded signal when loading contacts at startup
For the most part, this is done before any object has a chance to attach
to the contactAdded signal. Most components initialize by reading the
contacts list as well. But, other components (like FileTransferManager)
created from UserIdentity might try to do this before contacts are
loaded, so we need to emit the signal as well.