WIP Unify unix socket paths and work around issues with spaces
This is an attempt to work around issues with unix socket paths
containing spaces in Tor; but unfortunately, it's not enough. Tor will
refuse to place the control socket in a directory that isn't owned by
the user, unless we set a flag that doesn't exist until too-new versions
of Tor.